+++ python/aitemplate/frontend/nn/conv2d/conv2d_depthwise.py
@@ -29,7 +29,7 @@ class Conv2dDepthwise(Conv2d):
         padding=0,
         dilation=1,
         groups=1,
-        dtype="float16",
+        dtype="float32",
     ):
         super().__init__(
             in_channels,
